requirements.txt
pandas>=2.1
numpy>=1.26
pydantic>=2
jinja2>=3
rich
typer[all]
httpx
pyyaml
rapidfuzz
playwright
gradio
gradio-pdf